+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Bug fixes:
   * Fixed runtime issues in Java when a 'null' is returned from C++ for a non-nullable `Blob`.
+  * Updated Dart generator to generate `void` function return type explicitly.  
 
 ## 9.1.0
 Release date: 2021-05-31